ai-service/requirements.txt
flask
flask-cors
scikit-learn
numpy
psycopg2-binary
python-dotenv
